# docs/swagger.yaml
basePath: /
definitions:
  models.Capsule:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      topic:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  models.Topic:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Knowledge Capsule - knowledge management backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Knowledge Capsule
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: User login info
        in: body
        name: input
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email and password
      parameters:
      - description: User registration info
        in: body
        name: input
        required: true
        schema:
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/capsules:
    get:
      consumes:
      - application/json
      description: Get all capsules for the user (paginated) or create a new one
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Capsule info (for POST)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Capsule'
      produces:
      - application/json
      responses:
        "200":
          description: 'Paginated list: data, page, limit, total'
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Capsule'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get or create capsules
      tags:
      - capsules
    post:
      consumes:
      - application/json
      description: Get all capsules for the user (paginated) or create a new one
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Capsule info (for POST)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Capsule'
      produces:
      - application/json
      responses:
        "200":
          description: 'Paginated list: data, page, limit, total'
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Capsule'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get or create capsules
      tags:
      - capsules
  /api/capsules/{id}:
    delete:
      consumes:
      - application/json
      description: Get, update, or delete a single capsule (user must own it)
      parameters:
      - description: Capsule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated capsule fields (for PUT)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Capsule'
      produces:
      - application/json
      responses:
        "200":
          description: Capsule fetched or updated
          schema:
            $ref: '#/definitions/models.Capsule'
        "404":
          description: Capsule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get, update, or delete a capsule by ID
      tags:
      - capsules
    get:
      consumes:
      - application/json
      description: Get, update, or delete a single capsule (user must own it)
      parameters:
      - description: Capsule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated capsule fields (for PUT)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Capsule'
      produces:
      - application/json
      responses:
        "200":
          description: Capsule fetched or updated
          schema:
            $ref: '#/definitions/models.Capsule'
        "404":
          description: Capsule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get, update, or delete a capsule by ID
      tags:
      - capsules
    put:
      consumes:
      - application/json
      description: Get, update, or delete a single capsule (user must own it)
      parameters:
      - description: Capsule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated capsule fields (for PUT)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Capsule'
      produces:
      - application/json
      responses:
        "200":
          description: Capsule fetched or updated
          schema:
            $ref: '#/definitions/models.Capsule'
        "404":
          description: Capsule not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get, update, or delete a capsule by ID
      tags:
      - capsules
  /api/chat/history:
    get:
      consumes:
      - application/json
      description: Get paginated chat history between current user and another user
      parameters:
      - description: Other user ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Paginated messages: data, page, limit, total'
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get chat history
      tags:
      - chat
  /api/search:
    get:
      consumes:
      - application/json
      description: Search capsules by query string (paginated)
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Paginated list: data, page, limit, total'
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search capsules
      tags:
      - search
  /api/topics:
    get:
      consumes:
      - application/json
      description: Get all topics (paginated) or create a new one
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Topic info (for POST)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: 'Paginated list: data, page, limit, total'
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Topic'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get or create topics
      tags:
      - topics
    post:
      consumes:
      - application/json
      description: Get all topics (paginated) or create a new one
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Topic info (for POST)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: 'Paginated list: data, page, limit, total'
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Topic'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get or create topics
      tags:
      - topics
  /api/topics/{id}:
    delete:
      consumes:
      - application/json
      description: Get, update, or delete a single topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated topic fields (for PUT)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: Topic fetched or updated
          schema:
            $ref: '#/definitions/models.Topic'
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get, update, or delete a topic by ID
      tags:
      - topics
    get:
      consumes:
      - application/json
      description: Get, update, or delete a single topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated topic fields (for PUT)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: Topic fetched or updated
          schema:
            $ref: '#/definitions/models.Topic'
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get, update, or delete a topic by ID
      tags:
      - topics
    put:
      consumes:
      - application/json
      description: Get, update, or delete a single topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated topic fields (for PUT)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Topic'
      produces:
      - application/json
      responses:
        "200":
          description: Topic fetched or updated
          schema:
            $ref: '#/definitions/models.Topic'
        "404":
          description: Topic not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get, update, or delete a topic by ID
      tags:
      - topics
  /health:
    get:
      description: Check if the service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
swagger: "2.0"